backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
aiosqlite==0.19.0
pillow>=10.0.0
numpy>=1.24.0
PyJWT==2.8.0
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.15.0 
groq>=1.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0